Renames and organizes ttnn files
