+++ models/experimental/detr3d/ttnn/transformer_decoder.py
@@ -3,67 +3,8 @@
 # SPDX-License-Identifier: Apache-2.0
 
 import ttnn
-import math
 from models.common.lightweightmodule import LightweightModule
-
-
-class TTNNMultiheadAttention:
-    def __init__(self, d_model, nhead, device):
-        self.d_model = d_model
-        self.nhead = nhead
-        self.head_dim = d_model // nhead
-        self.device = device
-
-        # These will be set from PyTorch weights
-        self.q_weight = None
-        self.k_weight = None
-        self.v_weight = None
-        self.q_bias = None
-        self.k_bias = None
-        self.v_bias = None
-        self.out_weight = None
-        self.out_bias = None
-
-    def __call__(self, query, key, value, attn_mask=None):
-        # Apply linear projections separately to avoid concat issues
-        q = ttnn.linear(query, self.q_weight, bias=self.q_bias, memory_config=ttnn.DRAM_MEMORY_CONFIG)
-        k = ttnn.linear(key, self.k_weight, bias=self.k_bias, memory_config=ttnn.DRAM_MEMORY_CONFIG)
-        v = ttnn.linear(value, self.v_weight, bias=self.v_bias, memory_config=ttnn.DRAM_MEMORY_CONFIG)
-
-        # Get dimensions for reshaping
-        batch_size = q.shape[0]
-        q_seq_len = q.shape[1]
-        k_seq_len = k.shape[1]
-
-        # Reshape each tensor separately for multi-head attention
-        # [batch, seq_len, d_model] -> [batch, seq_len, num_heads, head_dim] -> [batch, num_heads, seq_len, head_dim]
-        q = ttnn.reshape(q, (batch_size, q_seq_len, self.nhead, self.head_dim))
-        q = ttnn.permute(q, (0, 2, 1, 3))
-
-        k = ttnn.reshape(k, (batch_size, k_seq_len, self.nhead, self.head_dim))
-        k = ttnn.permute(k, (0, 2, 1, 3))
-
-        v = ttnn.reshape(v, (batch_size, k_seq_len, self.nhead, self.head_dim))
-        v = ttnn.permute(v, (0, 2, 1, 3))
-
-        # Use SDPA for attention computation
-        context = ttnn.transformer.scaled_dot_product_attention(
-            q,
-            k,
-            v,
-            is_causal=False,
-            scale=1.0 / math.sqrt(self.head_dim),
-            attn_mask=attn_mask,
-            memory_config=ttnn.DRAM_MEMORY_CONFIG,
-        )
-
-        # Concatenate heads back to original format
-        context = ttnn.transformer.concatenate_heads(context, memory_config=ttnn.DRAM_MEMORY_CONFIG)
-
-        # Output projection
-        output = ttnn.linear(context, self.out_weight, bias=self.out_bias, memory_config=ttnn.DRAM_MEMORY_CONFIG)
-
-        return output
+from models.experimental.detr3d.ttnn.multihead_attention import TTNNMultiheadAttention
 
 
 class TTTransformerDecoderLayer(LightweightModule):
@@ -248,135 +189,6 @@ class TTTransformerDecoderLayer(LightweightModule):
         if return_attn_weights:
             return tgt, None
         return tgt, None
-
-
-class TTTransformerEncoderLayer(LightweightModule):
-    def __init__(
-        self,
-        device,
-        d_model,
-        nhead=4,
-        dim_feedforward=128,
-        normalize_before=True,
-        use_ffn=True,
-        parameters=None,
-    ):
-        super().__init__()
-        self.device = device
-        self.d_model = d_model
-        self.nhead = nhead
-        self.dim_feedforward = dim_feedforward
-        self.normalize_before = normalize_before
-        self.use_ffn = use_ffn
-
-        self.self_attn = TTNNMultiheadAttention(d_model, nhead, device)
-
-        # Load preprocessed parameters
-        if parameters is not None:
-            self.load_parameters(parameters)
-        else:
-            # Initialize weights as None (for backward compatibility)
-            self.self_attn_weights = None
-            self.ff_weights1 = None
-            self.ff_weights2 = None
-            self.norm1_weights = None
-            self.norm2_weights = None
-
-    def load_parameters(self, parameters):
-        """Load preprocessed parameters from the preprocessor"""
-        # Self-attention weights
-        if "self_attn" in parameters:
-            self.self_attn.q_weight = parameters["self_attn"].get("q_weight")
-            self.self_attn.k_weight = parameters["self_attn"].get("k_weight")
-            self.self_attn.v_weight = parameters["self_attn"].get("v_weight")
-            self.self_attn.q_bias = parameters["self_attn"].get("q_bias")
-            self.self_attn.k_bias = parameters["self_attn"].get("k_bias")
-            self.self_attn.v_bias = parameters["self_attn"].get("v_bias")
-            self.self_attn.out_weight = parameters["self_attn"].get("out_weight")
-            self.self_attn.out_bias = parameters["self_attn"].get("out_bias")
-
-        # Feedforward weights
-        if "linear1" in parameters:
-            self.ff_weights1 = parameters["linear1"]["weight"]
-            self.ff1_bias = parameters["linear1"].get("bias")
-        if "linear2" in parameters:
-            self.ff_weights2 = parameters["linear2"]["weight"]
-            self.ff2_bias = parameters["linear2"].get("bias")
-
-        # Normalization weights
-        for i, norm_name in enumerate(["norm1", "norm2"], 1):
-            if norm_name in parameters:
-                setattr(self, f"norm{i}_weights", parameters[norm_name]["weight"])
-                setattr(self, f"norm{i}_bias", parameters[norm_name].get("bias"))
-
-    def with_pos_embed(self, tensor, pos):
-        return tensor if pos is None else ttnn.add(tensor, pos)
-
-    def forward(
-        self,
-        src,
-        src_mask=None,
-        pos=None,
-        return_attn_weights=False,
-    ):
-        if self.normalize_before:
-            return self.forward_pre(src, src_mask, pos, return_attn_weights)
-        else:
-            return self.forward_post(src, src_mask, pos, return_attn_weights)
-
-    def forward_post(
-        self,
-        src,
-        src_mask=None,
-        pos=None,
-        return_attn_weights=False,
-    ):
-        q = k = self.with_pos_embed(src, pos)
-        value = src
-
-        # Self-attention
-        src2 = self.self_attn(q, k, value, src_mask)
-        src = ttnn.add(src, src2)
-        src = ttnn.layer_norm(src, weight=self.norm1_weights, bias=getattr(self, "norm1_bias", None))
-
-        # Feedforward network
-        if self.use_ffn:
-            src2 = ttnn.linear(src, self.ff_weights1, bias=getattr(self, "ff1_bias", None))
-            src2 = ttnn.relu(src2)
-            src2 = ttnn.linear(src2, self.ff_weights2, bias=getattr(self, "ff2_bias", None))
-            src = ttnn.add(src, src2)
-            src = ttnn.layer_norm(src, weight=self.norm2_weights, bias=getattr(self, "norm2_bias", None))
-
-        if return_attn_weights:
-            return src, None  # TTNN doesn't return attention weights
-        return src
-
-    def forward_pre(
-        self,
-        src,
-        src_mask=None,
-        pos=None,
-        return_attn_weights=False,
-    ):
-        src = ttnn.to_layout(src, ttnn.TILE_LAYOUT)
-        # Pre-norm self-attention
-        src2 = ttnn.layer_norm(src, weight=self.norm1_weights, bias=getattr(self, "norm1_bias", None))
-        value = src2
-        q = k = self.with_pos_embed(src2, pos)
-        src2 = self.self_attn(q, k, value, src_mask)
-        src = ttnn.add(src, src2)
-
-        # Pre-norm feedforward
-        if self.use_ffn:
-            src2 = ttnn.layer_norm(src, weight=self.norm2_weights, bias=getattr(self, "norm2_bias", None))
-            src2 = ttnn.linear(src2, self.ff_weights1, bias=getattr(self, "ff1_bias", None))
-            src2 = ttnn.relu(src2)
-            src2 = ttnn.linear(src2, self.ff_weights2, bias=getattr(self, "ff2_bias", None))
-            src = ttnn.add(src, src2)
-
-        if return_attn_weights:
-            return src, None
-        return src
 
 
 class TTTransformerDecoder(LightweightModule):
